feat(ma): calendar - add hook useYearSwitcher and change methods increase/decrease in service
