--- mother-frontend/micro-apps/calendar/src/ui/modules/year-switcher/use-year-switcher.composable.ts
@@ -1,25 +1,0 @@
-import useCalendar from '../../composables/use-calendar.composable'
-
-export default function useYearSwitcher() {
-  const { calendarService } = useCalendar()
-
-  const increaseYear = () => {
-    calendarService.increaseYear()
-
-    calendarService.changeCalendarByYear()
-  }
-
-  const decreaseYear = () => {
-    calendarService.decreaseYear()
-
-    calendarService.changeCalendarByYear()
-  }
-
-  const year = calendarService.year
-
-  return {
-    increaseYear,
-    decreaseYear,
-    year,
-  }
-}
