+++ mother-frontend/micro-apps/calendar/src/ui/modules/year-switcher/year-switcher.tsx
@@ -1,13 +1,16 @@
 import { observer } from 'mobx-react-lite'
 import { DSStepper } from '@windows98/design-system';
+import useYearSwitcher from './use-year-switcher.hook';
 
 export const YearSwitcher = observer(() => {
+	const { decreaseYear, increaseYear, year } = useYearSwitcher()
+
   	return (
 		<DSStepper
 			id="calendar-year"
-			modelValue={1900}
-			onIncrease={() => {}}
-			onDecrease={() => {}}
+			modelValue={year}
+			onIncrease={increaseYear}
+			onDecrease={decreaseYear}
 		/>
 	)
 })
